% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_optics_io.R
\name{ocean_lst2df}
\alias{ocean_lst2df}
\title{ocean list transform to data frame for use in ggplot() function}
\usage{
ocean_lst2df(ocean_lst)
}
\arguments{
\item{ocean_lst}{list of data read from multiple ocean spectrometer files}
}
\value{
ocean_df data.frame with transformed data (wavelength, I, name)
}
\description{
ocean list transform to data frame for use in ggplot() function
}
\examples{
\dontrun{ocean_lst()}
}
\author{
K. Juraic
}
