% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_run.R
\name{puma_edit_par}
\alias{puma_edit_par}
\title{edit set of parameters for call puma program}
\usage{
puma_edit_par(puma_par = NA, dat_file_name = "sample-dat.txt")
}
\arguments{
\item{puma_par}{list with parameters for call of puma program}

\item{dat_file_name}{name of file with experimental data ([file_name-dat.txt])}
}
\value{
list of puma parameters
}
\description{
edit set of parameters for call puma program
}
\examples{
     \dontrun{puma_edit_par()}
}
\author{
K. Juraic
}
