% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_io.R
\name{puma_read_inf}
\alias{puma_read_inf}
\title{read inf puma file with calculation results}
\usage{
puma_read_inf(datoteka)
}
\arguments{
\item{datoteka}{file name (-inf-dat) with results of puma calculations}
}
\value{
list with results
}
\description{
read inf puma file with calculation results
}
\examples{
     \dontrun{puma_read_inf(fname)}
}
