% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_run.R
\name{puma_run_iter}
\alias{puma_run_iter}
\title{run puma single iteration}
\usage{
puma_run_iter(puma_par, puma_path = "~/Job/R/Rpackages/rPuma/inst/puma")
}
\arguments{
\item{puma_par}{list of parameters required ro run puma progrm}

\item{puma_path}{path to puma program}
}
\value{
data_nk list of results of puma calculation
}
\description{
run puma single iteration
}
\examples{
      \dontrun{puma_run_iter(puma_par)}
}
