% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_optics_io.R
\name{read_ocean_txt}
\alias{read_ocean_txt}
\title{read ocean optics ASCII file with or without header}
\usage{
read_ocean_txt(file_name, header = TRUE)
}
\arguments{
\item{file_name}{Ocean Optics file name}

\item{header}{TRUE = file with header; FALSE = file without header}
}
\value{
list() with data
}
\description{
read ocean optics ASCII file with or without header
}
\examples{
     \dontrun{read_ocean_txt("ocean_file.txt")}
}
